mechanistic-interpretability/DPO/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
peft>=0.7.0
trl>=0.7.10
pyyaml>=6.0
wandb>=0.16.0
bitsandbytes>=0.41.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
ipywidgets>=8.1.0